make createConverterToAvro method public static so that it can be accessed outside of AvroOutputWriter
